Manually set throttle and record thrust
This modification to PICOLOv1 incorporates code from TestStand_Thrust and ESC_Controller which allows the flight computer to change the throttle percentage on the ESC motor through manual serial inputs. Both the throttle percentages and the applied thrust on the test stand are recorded in the table outputted to the serial monitor which can be read as a .csv file.