Propeller triggers at 2500ft
This modification to PICOLOv1 automates the propeller activation, to an extent. Instead of having to manually enter throttle percentages, the code automatically activates the propeller when the payload reaches 2500ft and sets the throttle to 50%. This is accomplished by comparing the value of the variable gpsAltFt to 2500. Once gpsAltFt equals 2500, the propeller is activated and spins for 10 seconds. Future modifications to this code will allow to the propeller to trigger once every 2500ft at different throttle values - which is how thrust data will be recorded.