Propeller switches off at 100,00ft
This modification to PICOLOv1 allows the propeller to switch on for 10 seconds at 50% throttle once every 2500ft. Additionally, a Boolean variable named AltMAX was introduced. When payload reaches 100,000ft, AltMAX would be set to true which indicated that the payload had reached a little under its hypothetical maximum/burst height. This tells the program to stop turning the propeller on as it may interfere with the payload landing and retrieval.