Incorporated LED panel
This modification to PICOLOv1 makes a large change in the propeller activation cycles. Now, during the propeller on time, the throttle decrements by 25% every 2 seconds from 100% to 25% - with each propeller activation lasting roughly 8 seconds. To indicate which throttle percentage the propeller is at, an LED panel is placed on the payload - in view of the camera. The panel has 4 LEDs, with each one corresponding to a different throttle value: Blue = 100%, Green = 75%, Yellow = 50%, Red = 25%, and no activated LEDs corresponds to the propeller being switched off (0% throttle).