Launch Edition
This modification to PICOLOv1 swaps the idea of propeller activation based on altitude variation with time variation. The balloon is assumed to ascend at an average rate of 1000 ft/min therefore, it can be assumed that the balloon ascends by 2500 ft roughly every 3 minutes. Instead of activating every 2500 ft, the propeller now activates every 3 minutes. The reason for this is that the flight computer may not read exactly 2500 ft when ascending and ground testing did not show much strong support for this either, despite adding a leeway of 15 ft. To ensure that the propeller does not activate until the balloon is launched, a 15 minute start-up delay was implemented - giving enough time to plug in the batteries, switch on the computer, strap up the payload, and tie the rigging.