Vacuum Chamber Modification
This modification to PICOLOv1 fixes a few bugs in the Launch Edition of the code where the taring would offset the recorded thrust to the negative range and made it difficult to see differences in thrust between different throttles. Now, the test stand is tared only during start up - allowing thrust to easily be differentiated at different throttles. Additionally, the throttle range was changed from 100 - 25 to 80 - 20 and decremented by 20% instead of 25%. This was done because it was later noticed that the ESC motor being used is typically meant to function within 20% to 60% throttle and should not be pushed beyond that. This version of the code was used specifically for post flight testing in a vacuum chamber.